Split service specific code out of main and remove global vars/funcs.
To prepare for making the code a bit more testable, this patch splits
the code that is not directly in main() out into its own files. In this
process, the antipattern of using global variables for cancel funcs etc
is being removed as well, and the service classes no longer call
os.Exit() themselves (it is done instead by the code in main).